Feature/surface pressure cpp operator integration (#3708)
* Changing SfcP to Sfc operator

* Hofx when missing values

* Fix the missing value check

* Fix the yaml file and fix the missing hofx value

* Fix the index to get operator object and output file name

* Add new file names and new rms ref value

* Missed the main yaml file to be changed

* This yaml should not have been changed, it will be removed later

* Change to create vector of missing value

* Update rms value, since the obs value of 1e+11 has been replaced with missingvalue. Geovals file name has been changed

* Empty-Commit to trigger CI

* Renaming the yaml file for consistency

* Integrate the yaml file name change in unit test

* Empty-Commit to start CI

---------

+++ src/ufo/operators/sfccorrected/EvalSurfacePressure.h
@@ -102,7 +102,7 @@ class stationPressure_GSL : public SurfaceOperatorBase {
                      std::vector<float> &,
                      std::vector<float> &,
                      std::vector<float> &,
-                     std::vector<double> &,
+                     std::vector<float> &,
                      std::vector<double> &,
                      std::vector<float> &) const;
 };
